feat: move auth confirm logic to host mfe
