+++ task-auth/src/routeTree.gen.ts
@@ -13,7 +13,6 @@
 import { Route as rootRoute } from './routes/__root'
 import { Route as SignUpImport } from './routes/sign-up'
 import { Route as LogInImport } from './routes/log-in'
-import { Route as AuthConfirmImport } from './routes/auth/confirm'
 
 // Create/Update Routes
 
@@ -24,11 +23,6 @@ const SignUpRoute = SignUpImport.update({
 
 const LogInRoute = LogInImport.update({
   path: '/log-in',
-  getParentRoute: () => rootRoute,
-} as any)
-
-const AuthConfirmRoute = AuthConfirmImport.update({
-  path: '/auth/confirm',
   getParentRoute: () => rootRoute,
 } as any)
 
@@ -50,23 +44,12 @@ declare module '@tanstack/react-router' {
       preLoaderRoute: typeof SignUpImport
       parentRoute: typeof rootRoute
     }
-    '/auth/confirm': {
-      id: '/auth/confirm'
-      path: '/auth/confirm'
-      fullPath: '/auth/confirm'
-      preLoaderRoute: typeof AuthConfirmImport
-      parentRoute: typeof rootRoute
-    }
   }
 }
 
 // Create and export the route tree
 
-export const routeTree = rootRoute.addChildren({
-  LogInRoute,
-  SignUpRoute,
-  AuthConfirmRoute,
-})
+export const routeTree = rootRoute.addChildren({ LogInRoute, SignUpRoute })
 
 /* prettier-ignore-end */
 
@@ -77,8 +60,7 @@ export const routeTree = rootRoute.addChildren({
       "filePath": "__root.tsx",
       "children": [
         "/log-in",
-        "/sign-up",
-        "/auth/confirm"
+        "/sign-up"
       ]
     },
     "/log-in": {
@@ -86,9 +68,6 @@ export const routeTree = rootRoute.addChildren({
     },
     "/sign-up": {
       "filePath": "sign-up.tsx"
-    },
-    "/auth/confirm": {
-      "filePath": "auth/confirm.tsx"
     }
   }
 }
