+++ task-auth/src/services/auth/index.ts
@@ -1,6 +1,6 @@
 import { SignInWithPasswordCredentials } from '@supabase/supabase-js';
 import { db } from '../db';
-import { AuthConfirmQueryParams, AuthData } from './types';
+import { AuthData } from './types';
 
 export const logIn = async (data: AuthData) => {
     const { error } = await db.auth.signInWithPassword(data as unknown as SignInWithPasswordCredentials);
@@ -16,20 +16,4 @@ export const signUp = async (data: AuthData) => {
     if (error) {
         throw error;
     }
-};
-
-export const authConfirm = async ({ token_hash, type }: AuthConfirmQueryParams) => {
-    if (token_hash && type) {
-        const { error } = await db.auth.verifyOtp({
-            type,
-            token_hash,
-        });
-
-        if (!error) {
-            location.replace('/');
-            return;
-        }
-    }
-
-    location.replace('/error');
 };
