+++ task-auth/src/services/auth/types.ts
@@ -1,11 +1,4 @@
-import { EmailOtpType } from '@supabase/supabase-js';
-
 export interface AuthData {
     email: string;
     password: string;
-}
-
-export interface AuthConfirmQueryParams {
-    token_hash: string;
-    type: EmailOtpType | null;
 }
