+++ task-host/src/services/auth/types.ts
@@ -1,0 +1,6 @@
+import { EmailOtpType } from '@supabase/supabase-js';
+
+export interface AuthConfirmQueryParams {
+    token_hash: string;
+    type: EmailOtpType;
+}
